# api/finance_openapi.yaml
openapi: 3.0.3
info:
  title: Frego Finance API
  description: Finance module API for Frego ERP - handles AR, AP, and GL operations
  version: 1.0.0
  contact:
    name: Frego Support
    email: support@frego.com
servers:
  - url: http://localhost:8080/finance/api/v1
    description: Local development
  - url: https://api.myfrego.com/finance/api/v1
    description: Production
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
...
paths:
...
  /tenants/provision:
    post:
      summary: Provision finance schema for a tenant
      operationId: provisionTenant
      description: Creates finance schema for the tenant. Schema name is computed from displayName with 'fin_' prefix, or generated from tenantId if displayName is not provided.
      security: []
      parameters:
        - name: Secret
          in: header
          required: true
          description: Internal secret used for provisioning authorization
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tenantId
              properties:
                tenantId:
                  type: string
                  format: uuid
                  description: Tenant identifier
                displayName:
                  type: string
                  description: Tenant display name (used to compute schema name)
                actor:
                  type: string
                  description: User/actor performing the provisioning
      responses:
        '200':
          description: Tenant provisioned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  tenantId:
                    type: string
                    format: uuid
                  schemaName:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'